requirements.txt
python3-flask=2.2.2
python3-flask-socketio=5.3.2
python3-flask-cors=3.0.10
python3-psutil=5.9.4
python3-picamera2=0.3.21
python3-libcamera
python3-eventlet=0.33.1
python3-twisted=22.4.0
python3-cryptography=38.0.4
python3-opencv
python3-serial
xdg-utils